Move related functions into the same modules.
